requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-cors==0.8.1
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==3.7.1
attrs==25.3.0
black==23.11.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorful==0.5.7
distlib==0.3.9
fastapi==0.104.1
filelock==3.18.0
flake8==6.1.0
frozenlist==1.7.0
fsspec==2025.5.1
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.73.1
h11==0.16.0
httptools==0.6.4
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mccabe==0.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
mypy_extensions==1.1.0
networkx==3.5
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencensus==0.11.4
opencensus-context==0.1.3
opentelemetry-api==1.34.1
opentelemetry-exporter-prometheus==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
prometheus_client==0.22.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
py-spy==0.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyflakes==3.1.0
pytest==7.4.3
python-dotenv==1.1.1
python-multipart==0.0.6
PyYAML==6.0.2
pyzmq==25.1.1
ray==2.47.1
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
rsa==4.9.1
setuptools==78.1.1
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
starlette==0.27.0
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
virtualenv==20.31.2
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
